% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisaL.R
\docType{data}
\name{pisaL}
\alias{pisaL}
\title{PISA example data with Responses and Response Times}
\format{
A data frame with xx rows and xx variables
}
\usage{
data(pisaL)
}
\description{
A data set from the Canadian subsample of the PISA 2015 assessment. Containts respones and response times of a single math booklet.
}
\details{
\itemize{
  \item responses
  \item response times
}
}
\keyword{datasets}
